backend-graph/requirements.txt
Django==4.2
djangorestframework==3.14.0
matplotlib==3.7.1
pandas==2.0.2
numpy==1.26.4
uvicorn==0.22.0